commit-tree.c: don't append a newline with -F
This change makes it such that commit-tree -F never appends a newline
character to the supplied commit message (either from file or stdin).

Previously, commit-tree -F would always append a newline character to
the text brought in from file or stdin. This has caused confusion in a
number of ways:
1. This is a plumbing command and it is generally expected not to do
   text cleanup or other niceties.
2. stdin piping with "-F -" appends a newline but stdin piping without
   -F does not append a newline (inconsistent).
3. git-commit has the --cleanup=verbatim option that prevents appending
   a newline with its -F argument. There is no verbatim counterpart to
   commit-tree -F (inconsistent).